+++ TablePro/Views/Structure/TableStructureView.swift
@@ -178,6 +178,8 @@ struct TableStructureView: View {
     private var contentArea: some View {
         if let error = errorMessage {
             errorView(error)
+        } else if isInitialLoading {
+            DataGridSkeletonView()
         } else {
             tabContent
         }